+++ src/models/index.ts
@@ -2,6 +2,8 @@ export * from './MetaModel';
 
 export * from './SubStoreModel';
 
+export * from './lists';
+
 export * from './timers';
 
 export * from './BlurAndFocusHandlerModel';