+++ src/models/lists/ApiListModel/index.ts
@@ -1,0 +1,3 @@
+export { default as ApiListModel } from './ApiListModel';
+
+export * from './types';