+++ src/models/lists/MockApiListModel/index.ts
@@ -1,0 +1,3 @@
+export { default as MockApiListModel } from './MockApiListModel';
+
+export * from './types';